requirements.txt
pandas>=2.2.0
numpy>=1.26.0
pydantic>=2.7.0
PyYAML>=6.0.1
requests>=2.31.0
python-dotenv>=1.0.1
pytest>=8.0.0
